# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Ticket Booking System API
  description: |
    A comprehensive REST API for managing movie ticket bookings, theatres, shows, and user authentication.

    This API provides:
    - User authentication and authorization
    - Movie and theatre management
    - Show scheduling and seat management
    - Transaction-safe booking system with stored procedures
    - Comprehensive error handling and validation

    **Key Features:**
    - JWT-based authentication
    - Role-based access control (USER, ADMIN)
    - Real-time seat availability
    - Secure payment processing
    - Database-first architecture with MySQL
  version: 1.0.0
  contact:
    name: Movie Booking System Support
    email: support@moviebooking.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.moviebooking.com/api
    description: Production server
paths:
...
  /movies:
    get:
      tags:
        - Movies
      summary: List movies
      description: Retrieve a paginated list of movies with optional filtering
      operationId: getMovies
      parameters:
        - $ref: '#/components/parameters/GenreFilter'
        - $ref: '#/components/parameters/LanguageFilter'
        - $ref: '#/components/parameters/RatingFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Movies retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviesResponse'

    post:
      tags:
        - Movies
      summary: Create movie (Admin only)
      description: Create a new movie entry
      operationId: createMovie
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '201':
          description: Movie created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /movies/search:
    get:
      tags:
        - Movies
      summary: Search movies
      description: Search movies by title, description, or other criteria
      operationId: searchMovies
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
            maxLength: 100
          example: "batman"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviesResponse'
  /movies/{id}:
    get:
      tags:
        - Movies
      summary: Get movie details
      description: Retrieve detailed information about a specific movie
      operationId: getMovie
      parameters:
        - $ref: '#/components/parameters/MovieId'
      responses:
        '200':
          description: Movie details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Movies
      summary: Update movie (Admin only)
      description: Update an existing movie
      operationId: updateMovie
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MovieId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '200':
          description: Movie updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Movies
      summary: Delete movie (Admin only)
      description: Soft delete a movie (sets is_active to false)
      operationId: deleteMovie
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MovieId'
      responses:
        '200':
          description: Movie deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Theatre Endpoints
  /theatres:
    get:
      tags:
        - Theatres
      summary: List theatres
      description: Retrieve a list of all active theatres
      operationId: getTheatres
      responses:
        '200':
          description: Theatres retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheatresResponse'

    post:
      tags:
        - Theatres
      summary: Create theatre (Admin only)
      description: Create a new theatre
      operationId: createTheatre
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TheatreInput'
      responses:
        '201':
          description: Theatre created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheatreResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /theatres/{id}:
    get:
      tags:
        - Theatres
      summary: Get theatre details
      description: Retrieve detailed information about a theatre including screens
      operationId: getTheatre
      parameters:
        - $ref: '#/components/parameters/TheatreId'
      responses:
        '200':
          description: Theatre details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheatreDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Show Endpoints
  /shows:
    get:
      tags:
        - Shows
      summary: List shows
      description: Retrieve shows with optional filtering by movie, theatre, or date
      operationId: getShows
      parameters:
        - name: movieId
          in: query
          description: Filter by movie ID
          schema:
            type: integer
            minimum: 1
        - name: theatreId
          in: query
          description: Filter by theatre ID
          schema:
            type: integer
            minimum: 1
        - name: date
          in: query
          description: Filter by specific date
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          description: Filter from start date
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter to end date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Shows retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowsResponse'

    post:
      tags:
        - Shows
      summary: Create show (Admin only)
      description: Create a new show
      operationId: createShow
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowInput'
      responses:
        '201':
          description: Show created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /shows/upcoming:
    get:
      tags:
        - Shows
      summary: Get upcoming shows
      description: Retrieve upcoming shows for the next few days
      operationId: getUpcomingShows
      parameters:
        - name: days
          in: query
          description: Number of days to look ahead (1-90)
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
      responses:
        '200':
          description: Upcoming shows retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowsResponse'
  /shows/{id}:
    get:
      tags:
        - Shows
      summary: Get show details
      description: Retrieve detailed information about a specific show
      operationId: getShow
      parameters:
        - $ref: '#/components/parameters/ShowId'
      responses:
        '200':
          description: Show details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /shows/{id}/seats:
    get:
      tags:
        - Shows
      summary: Get available seats
      description: |
        Retrieve seat availability for a specific show.
        This endpoint calls a stored procedure for real-time accuracy.
      operationId: getAvailableSeats
      parameters:
        - $ref: '#/components/parameters/ShowId'
      responses:
        '200':
          description: Seat availability retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatAvailabilityResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Booking Endpoints (CRITICAL)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    MovieId:
      name: id
      in: path
      required: true
      description: Movie ID
      schema:
        type: integer
        minimum: 1
    TheatreId:
      name: id
      in: path
      required: true
      description: Theatre ID
      schema:
        type: integer
        minimum: 1
    ShowId:
      name: id
      in: path
      required: true
      description: Show ID
      schema:
        type: integer
        minimum: 1
...
    GenreFilter:
      name: genre
      in: query
      description: Filter by genre
      schema:
        $ref: '#/components/schemas/Genre'
    LanguageFilter:
      name: language
      in: query
      description: Filter by language
      schema:
        type: string
        minLength: 2
        maxLength: 50
    RatingFilter:
      name: rating
      in: query
      description: Filter by rating
      schema:
        $ref: '#/components/schemas/Rating'
    Limit:
      name: limit
      in: query
      description: Number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
...
    Genre:
      type: string
      enum:
        - ACTION
        - COMEDY
        - DRAMA
        - HORROR
        - ROMANCE
        - SCI_FI
        - THRILLER
        - DOCUMENTARY
        - ANIMATION
        - OTHER
    Rating:
      type: string
      enum:
        - G
        - PG
        - PG-13
        - R
        - NC-17
        - NR
    Movie:
      type: object
      properties:
        id:
          type: integer
          description: Movie ID
        title:
          type: string
          description: Movie title
        description:
          type: string
          description: Movie description
        genre:
          $ref: '#/components/schemas/Genre'
        language:
          type: string
          description: Movie language
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 500
          description: Movie duration in minutes
        release_date:
          type: string
          format: date
          description: Release date
        rating:
          $ref: '#/components/schemas/Rating'
        poster_url:
          type: string
          format: uri
          description: Poster image URL
        is_active:
          type: boolean
          description: Whether movie is active
        created_at:
          type: string
          format: date-time
    MovieInput:
      type: object
      required:
        - title
        - description
        - genre
        - language
        - duration_minutes
        - release_date
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
        genre:
          $ref: '#/components/schemas/Genre'
        language:
          type: string
          minLength: 2
          maxLength: 50
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 500
        release_date:
          type: string
          format: date
        rating:
          $ref: '#/components/schemas/Rating'
        poster_url:
          type: string
          format: uri
    MovieResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            movie:
              $ref: '#/components/schemas/Movie'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    MoviesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            movies:
              type: array
              items:
                $ref: '#/components/schemas/Movie'
            pagination:
              $ref: '#/components/schemas/Pagination'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'

    # Theatre Schemas
    Theatre:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
          pattern: '^[0-9]{6}$'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    TheatreInput:
      type: object
      required:
        - name
        - location
        - city
        - state
        - pincode
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
        location:
          type: string
          minLength: 5
          maxLength: 255
        city:
          type: string
          minLength: 2
          maxLength: 100
        state:
          type: string
          minLength: 2
          maxLength: 100
        pincode:
          type: string
          pattern: '^[0-9]{6}$'
...
    TheatreResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            theatre:
              $ref: '#/components/schemas/Theatre'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    TheatreDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            theatre:
              $ref: '#/components/schemas/TheatreDetail'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    TheatresResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            theatres:
              type: array
              items:
                $ref: '#/components/schemas/Theatre'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'

    # Show Schemas
...
    ShowInput:
      type: object
      required:
        - movie_id
        - screen_id
        - show_time
        - price
      properties:
        movie_id:
          type: integer
          minimum: 1
        screen_id:
          type: integer
          minimum: 1
        show_time:
          type: string
          format: date-time
        price:
          type: number
          format: decimal
          minimum: 0.01
          multipleOf: 0.01
...
    ShowResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            show:
              $ref: '#/components/schemas/Show'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    ShowDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            show:
              $ref: '#/components/schemas/ShowDetail'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    ShowsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            shows:
              type: array
              items:
                $ref: '#/components/schemas/Show'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'

    # Seat Schemas
...
    SeatAvailabilityResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            show:
              type: object
              properties:
                id:
                  type: integer
                show_time:
                  type: string
                  format: date-time
                price:
                  type: number
                  format: decimal
            seats:
              type: array
              items:
                $ref: '#/components/schemas/Seat'
            summary:
              $ref: '#/components/schemas/SeatSummary'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'

    # Booking Schemas
...
    MetaData:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        api_version:
          type: string
          example: "1.0"
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        hasNext:
          type: boolean
          description: Whether there's a next page
        hasPrev:
          type: boolean
          description: Whether there's a previous page
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human readable error message
            details:
              type: array
              items:
                type: string
              description: Additional error details
            path:
              type: string
              description: API endpoint path
        meta:
          $ref: '#/components/schemas/MetaData'
  responses:
...
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTHENTICATION_ERROR"
              message: "Authentication required"
              path: "/api/bookings"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTHORIZATION_ERROR"
              message: "Admin access required"
              path: "/api/movies"
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "Movie not found"
              path: "/api/movies/999"
...
    ValidationError:
      description: Validation Error - Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Validation failed"
              details: ["Password must contain uppercase letter"]
              path: "/api/auth/register"
security:
  - bearerAuth: []
...